separate types file
