+++ utopia-remix/app/util/github-branch-contents.server.spec.ts
@@ -1,6 +1,4 @@
 import urlJoin from 'url-join'
-import type { ProjectContentDirectory } from '../types'
-import { projectContentDirectory, type ExistingAsset } from '../types'
 import type { AssetToUpload, UnzipEntry } from './github-branch-contents.server'
 import {
   populateDirectories,
@@ -12,6 +10,9 @@ import {
 import * as fs from 'fs'
 import * as os from 'os'
 import path from 'path'
+import type { ExistingAsset } from '../types'
+import type { ProjectContentDirectory } from '../types-project-contents'
+import { projectContentDirectory } from '../types-project-contents'
 
 describe('Github get branch contents', () => {
   describe('shouldUploadAsset', () => {
